A lot of basic clean up
Improve folder structure.
Remove unused code.
Move test-only code from production to test code.
