+++ source/Energinet.Charges.Libraries/source/Energinet.DataHub.Charges.Clients.IntegrationTests/DefaultChargeLink/DefaultChargeLinkTests.cs
@@ -16,11 +16,10 @@ using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
 using Azure.Messaging.ServiceBus;
-using Energinet.DataHub.Charges.Clients.Common;
 using Energinet.DataHub.Charges.Clients.DefaultChargeLink;
+using Energinet.DataHub.Charges.Clients.DefaultChargeLink.Models;
 using Energinet.DataHub.Charges.Clients.IntegrationTests.Fixtures;
-using Energinet.DataHub.Charges.Clients.Models;
-using Energinet.DataHub.Charges.Clients.Providers;
+using Energinet.DataHub.Charges.Clients.ServiceBus.Providers;
 using FluentAssertions;
 using GreenEnergyHub.TestHelpers;
 using Xunit;
