+++ source/Energinet.Charges.Libraries/source/Energinet.DataHub.Charges.Clients.Tests/DefaultChargeLink/DefaultChargeLinkClientTests.cs
@@ -17,9 +17,9 @@ using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
 using AutoFixture.Xunit2;
 using Energinet.DataHub.Charges.Clients.DefaultChargeLink;
-using Energinet.DataHub.Charges.Clients.Models;
-using Energinet.DataHub.Charges.Clients.Providers;
+using Energinet.DataHub.Charges.Clients.DefaultChargeLink.Models;
 using Energinet.DataHub.Charges.Clients.ServiceBus;
+using Energinet.DataHub.Charges.Clients.ServiceBus.Providers;
 using GreenEnergyHub.Charges.TestCore.Attributes;
 using GreenEnergyHub.TestHelpers;
 using Moq;
