+++ source/Energinet.Charges.Libraries/source/Energinet.DataHub.Charges.Clients/DefaultChargeLink/Models/DefaultChargeLinksCreatedSuccessfullyDto.cs
@@ -12,11 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Charges.Clients.Models
+namespace Energinet.DataHub.Charges.Clients.DefaultChargeLink.Models
 {
     /// <summary>
-    /// The required data needed by the Charges domain to create Default charge links.
+    /// The data needed by the Metering Point domain as a reply
+    /// to a successful CreateDefaultChargeLinks request
     /// </summary>
     /// <param name="MeteringPointId">A unique id to specify the metering point.</param>
-    public sealed record RequestDefaultChargeLinksForMeteringPointDto(string MeteringPointId);
+    /// <param name="DidCreateChargeLinks">True if any charge links was created.</param>
+    public sealed record DefaultChargeLinksCreatedSuccessfullyDto(string MeteringPointId, bool DidCreateChargeLinks);
 }
