+++ source/Energinet.Charges.Libraries/source/Energinet.DataHub.Charges.Clients/Energinet.DataHub.Charges.Clients.csproj
@@ -28,7 +28,7 @@ limitations under the License.
   <!-- Configuration for NuGet package -->
   <PropertyGroup>
     <PackageId>Energinet.DataHub.Charges.Clients</PackageId>
-    <PackageVersion>1.0.16$(VersionSuffix)</PackageVersion>
+    <PackageVersion>1.0.17$(VersionSuffix)</PackageVersion>
     <Title>Energinet.DataHub.Charges.Clients</Title>
     <Company>Energinet DataHub A/S</Company>
     <Authors>DataHub</Authors>
@@ -105,6 +105,12 @@ https://github.com/Energinet-DataHub/geh-charges/tree/main/source/Energinet.Char
       <OutputDir>obj\contracts\</OutputDir>
       <Generator>MSBuild:Compile</Generator>
     </Protobuf>
+    <Protobuf Update="..\..\..\Contracts\CreateDefaultChargeLinks.proto">
+      <Link>DefaultChargeLink\Contracts\CreateDefaultChargeLinks.proto</Link>
+    </Protobuf>
+    <Protobuf Update="..\..\..\Contracts\CreateDefaultChargeLinksReply.proto">
+      <Link>DefaultChargeLink\Contracts\CreateDefaultChargeLinksReply.proto</Link>
+    </Protobuf>
   </ItemGroup>
   
   <ItemGroup>
@@ -131,10 +137,14 @@ https://github.com/Energinet-DataHub/geh-charges/tree/main/source/Energinet.Char
   </ItemGroup>
 
   <ItemGroup>
-    <Compile Include="..\..\..\Contracts\ChargeLinkDto.cs" Link="ChargeLinkDto.cs" />
-    <Compile Include="..\..\..\Contracts\ChargeType.cs" Link="ChargeType.cs" />
+    <Compile Include="..\..\..\Contracts\ChargeLinkDto.cs">
+      <Link>ChargeLinks\ChargeLinkDto.cs</Link>
+    </Compile>
+    <Compile Include="..\..\..\Contracts\ChargeType.cs">
+      <Link>ChargeLinks\ChargeType.cs</Link>
+    </Compile>
     <Compile Include="..\..\..\Contracts\ErrorCode.cs">
-      <Link>Common\ErrorCode.cs</Link>
+      <Link>DefaultChargeLink\ErrorCode.cs</Link>
     </Compile>
   </ItemGroup>
