+++ source/Energinet.Charges.Libraries/source/Energinet.DataHub.Charges.Clients/documents/release-notes/release-notes.md
@@ -16,6 +16,10 @@ Updated NuGet dependencies and removed deprecated FxCopAnalyzers
 
 Removed `IDefaultChargeLinkReplyReader`
 
+## Version 1.0.13
+
+Internal restructuring.
+
 ## Version 1.0.12
 
 Updated a number of NuGet dependencies.
